tests for font size

+++ hackspace labels/makeLabel.py
@@ -1,11 +1,14 @@
 #!/usr/bin/python
 """
+another good font?
+http://www.fontsaddict.com/download/silexstencil-basic.otf
 program to make laser cutterable labels for bristol hackspace storage boxes.
 needs digital-7 font installed: http://www.dafont.com/digital-7.font
 relies on inkscape to make the svg into an eps, and then pstoedit to turn the file into a dxf. The magic argument to include is -mm, which gets the scale right.
 
 """
 import argparse
+from PIL import Image, ImageDraw, ImageFont
 from pysvg.shape import *
 from pysvg import parser
 from pysvg.style import *
@@ -15,21 +18,35 @@ from pysvg.builders import *
 def get_style():
     style=StyleBuilder()
     style.setFontFamily(fontfamily="Digital-7")
-    style.setFontSize(fontsize) 
+#    style.setFontFamily(fontfamily="FreeSans")
+    style.setFontSize(args.fontsize/3.39) 
     style.setFilling("black")
     return style.getStyle()
 
 def write_label(svg):
-    textWidth=len(args.text)*charWidth
+    (textWidth,textHeight)=get_font_size(fontsize)
+    print textWidth,textHeight
     if textWidth > width:
       print "text too long"
       exit(1)
-    x=(width-textWidth)/2
-    y=(height-charHeight)/2+charHeight
+    x=0 #(width-textWidth)/2
+    y=height #10 #height #(height-textHeight)/2+textHeight
+    margin = 0
     t=text(args.text,x+margin,y+margin)
     t.set_style(get_style())
     svg.addElement(t)
 
+#takes a font size and returns mm it is
+def get_font_size(fontsize):
+  font = ImageFont.truetype(args.font,args.fontsize) #fontsize)
+  im = Image.new('RGBA', (400,200),"yellow")
+  draw = ImageDraw.Draw(im)
+  w, h = font.getsize(args.text)
+  draw.text(((400-w)/2,(200-h)/2), args.text, fill="black",font=font)
+  im.save("hello.png", "PNG")
+  print w,h
+  print w/3.8,h/3.8
+  return(w/3.8, h/3.8)
 
 def square(svg):
   x=margin
@@ -58,7 +75,7 @@ def setup():
 
 if __name__ == '__main__':
   argparser = argparse.ArgumentParser(
-      description="generates square based energy drawings")
+      description="generates a laser cutterable label")
   argparser.add_argument('--file',
       action='store', dest='file', default="label.dxf",
       help="dxf file to create")
@@ -68,20 +85,25 @@ if __name__ == '__main__':
   argparser.add_argument('--size',
       action='store', dest='size', type=int, default=0,
       help="size, 0=small box, 1=big box")
+  argparser.add_argument('--fontsize',
+      action='store', type=int, dest='fontsize', default=None,
+      help="override font")
+  argparser.add_argument('--font',
+      action='store', dest='font', default=None,
+      help="specify where the font is installed")
+  argparser.add_argument('--noremove',
+      action='store_const', const=False, dest='remove', default=True,
+      help="don't remove temporary files")
 
   args = argparser.parse_args()
 
   if args.size == 0:
-    height=23
-    width=75
-    charWidth=4.5
-    charHeight=7
+    height=50 #23
+    width=100 #75
     fontsize=10 #results in text char size 5x7mm
   elif args.size == 1:
     height=35
     width=104
-    charWidth=6.8
-    charHeight=9.7
     fontsize=15 #results in text char size 5x7mm
 
   margin = 5
@@ -95,10 +117,11 @@ if __name__ == '__main__':
 
   import os
   #magic!
-  os.system("inkscape -E label.eps label.svg") 
-  os.system("pstoedit -dt -f dxf:'-polyaslines -mm' label.eps " + args.file)
+  #os.system("inkscape -E label.eps label.svg") 
+  #os.system("pstoedit -dt -f dxf:'-polyaslines -mm' label.eps " + args.file)
   #get rid of old temp files
-  os.system("rm label.svg")
-  os.system("rm label.eps")
+  if args.remove:
+    os.system("rm label.svg")
+    os.system("rm label.eps")
 
   print "laser cutter settings. power at 1 turn, 50mm per second"
